refactor: Move ZoomDetectorView class to animations package
Refactored the ZoomDetectorView class by moving it from its previous location to the animations package for better organization and clarity.

Files modified:
- Moved ZoomDetectorView class to animations package
